complete scheduled (need add method for second update state of report need fix writing dates to table)

+++ src/main/java/com/bot/employeeTimeTracongBot/model/User.java
@@ -19,9 +19,9 @@ public class User {
     private String nickName;
     private String fullName;
     private String dateLastReport;
-    private int hours;
+    private double hours;
 
-    public User(boolean access, boolean isSendReport, String name, String dateConnecting, int chatId, String nickName, String fullName, String dateLastReport, int hours) {
+    public User(boolean access, boolean isSendReport, String name, String dateConnecting, int chatId, String nickName, String fullName, String dateLastReport, double hours) {
         this.access = access;
         this.isSendReport = isSendReport;
         this.name = name;
